수정

+++ src/com/javaex/Problem02.java
@@ -10,8 +10,8 @@ public class Problem02 {
 
 		//다섯개의 숫자를 입력받아 평균을 구하시오
 		
-		int [] num = new int[5];
-		int total = 0;
+		long [] num = new long[5];
+		double total = 0;
 		
 		Scanner sc = new Scanner(System.in);
 		
@@ -20,7 +20,7 @@ public class Problem02 {
 		for(int i=0; i<num.length; i++){
 			try {
 				System.out.println("숫자를 입력하세요");
-				num[i] = sc.nextInt();
+				num[i] = sc.nextLong();
 				sc.nextLine();
 				total += num[i];
 			} catch (InputMismatchException e) {
@@ -30,7 +30,7 @@ public class Problem02 {
 				i--;
 			}
 		}
-		
+
 		System.out.println("평균은 : " + total/num.length);
 	}
